requirements.txt
flask
gunicorn
python-dotenv
psycopg2-binary
flask-cors
werkzeug
pillow
ffmpeg-python
bcrypt
flask-mail